added help for say command

+++ commands/admin/say.js
@@ -3,12 +3,15 @@ exports.run = async (client, message, args) => {
   const config = index.config
   const Discord = index.Discord
 
+  if(!args[0])
+    return this.help(client, message, args)
+
   if(!config.admins.ids[message.author.id]) {
     message.reply("no permission")
     return
   }
 
-  await message.channel.send(args.join(" ") || `you need to specify something for me to say dumbash ${message.author}`)
+  await message.channel.send(args.join(" "))
 
   if(message.deletable)
     message.delete()
@@ -16,12 +19,17 @@ exports.run = async (client, message, args) => {
 
 exports.help = async (client, message, args) => {
   const index = require("../../index.js")
-  const config = index.config
-  const Discord = index.Discord
+  const { commandHelpEmbed, prefix } = index
 
-  const emb = new Discord.MessageEmbed()
-    .setColor(config.main.colours.help)
-    .setTitle("say")
-    .setDescription("say")
-  return message.channel.send(emb)
+  const embed = commandHelpEmbed(message, {
+    title: "**Say Command**",
+    description: "Make g9lBot say something! (You must be listed as an admin to use this.)",
+    syntax: `${prefix}say <something>`,
+    example: [
+      `**say hi**`,
+      ` ${prefix}say hi`,
+    ].join("\n"),
+  })
+  
+  message.channel.send(embed)
 }
